requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.7.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pywebview>=5.0
